Section 9 Complete

+++ backend/routes/posts.js
@@ -34,7 +34,8 @@ router.post("", checkAuth, multer({ storage: storage }).single("image"), (req, r
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   })
   post.save().then(createdPost => {
     res.status(201).json({
@@ -58,11 +59,16 @@ router.put("/:id", checkAuth, multer({ storage: storage }).single("image"), (req
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   })
   //console.log(post)
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    res.status(200).json({ message: 'Update successful!' })
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({ message: 'Update successful!' })
+    } else {
+      res.status(401).json({ message: 'Not Authorized!' })
+    }
   })
 })
 
@@ -112,9 +118,14 @@ router.get("/:id", (req, res, next) => {
 
 router.delete("/:id", checkAuth, (req, res, next) => {
   //console.log(req.params.id)
-  Post.deleteOne({ _id: req.params.id }).then(result => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
     //console.log(result)
-    res.status(200).json({ message: 'Post deleted!' })
+    if (result.n > 0) {
+      res.status(200).json({ message: 'Post Deleted!' })
+    } else {
+      res.status(401).json({ message: 'Not Authorized to Delete Post!' })
+    }
+
   })
 
 });
